requirement.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2